backend2/requirements.txt
faiss-cpu==1.7.4
fastapi==0.115.6
uvicorn==0.32.1
PyMuPDF==1.24.9
numpy==1.26.4
Pillow==11.0.0
pydantic==2.10.4
pytesseract==0.3.13
sentence-transformers==3.3.1
torch==2.5.1+cpu
transformers==4.47.1
underthesea==6.8.4
python-multipart==0.0.12
aiofiles==24.1.0
